+++ index.html
@@ -10,8 +10,7 @@
 </head>
 
 <body>
-  <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
